initial debug of BensBoard and checkUtil
